Fix faction selection on object creation
The misc_object listbox generates scripts for each button when the
dialog opens, so the faction effectively cannot be changed in the
dialog. The workaround is to inject the faction into the script when the
button is clicked.

A cleaner fix would be to refactor the listbox or create a control which
is a scrollable list of GuiButton, but that seems like a change with
bigger impact.